+++ manager.cpp
@@ -450,7 +450,7 @@ cache_factory *manager::get_cache_factory()
 #endif
 #ifdef EN_TCP_CACHE
 	else if(backend=="tcp") {
-		vector<long> const &ports=config.llist("cache.ports");
+		vector<int> const &ports=config.llist("cache.ports");
 		vector<string> const &ips=config.slist("cache.ips");
 		return new tcp_cache_factory(ips,ports);
 	}
